Модернизация приложения
