+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/AndroidStudioProjects/Android-2022/Android-1/Lesson-1/MyFragmentsLessons/app/src/main/res/drawable/ic_baseline_arrow_back_24.xml" value="0.163" />
+        <entry key="..\:/AndroidStudioProjects/Android-2022/Android-1/Lesson-1/MyFragmentsLessons/app/src/main/res/drawable/ic_divider.xml" value="0.12" />
         <entry key="..\:/AndroidStudioProjects/Android-2022/Android-1/Lesson-1/MyFragmentsLessons/app/src/main/res/drawable/ic_launcher_background.xml" value="0.163" />
         <entry key="..\:/AndroidStudioProjects/Android-2022/Android-1/Lesson-1/MyFragmentsLessons/app/src/main/res/drawable/ic_launcher_foreground.xml" value="0.163" />
         <entry key="..\:/AndroidStudioProjects/Android-2022/Android-1/Lesson-1/MyFragmentsLessons/app/src/main/res/layout-land/activity_main.xml" value="0.1" />
