+++ app/src/main/java/come/geekbrains/myfragmentslessons/ui/MainActivity.java
@@ -14,7 +14,7 @@ import com.google.android.material.navigation.NavigationView;
 
 import come.geekbrains.myfragmentslessons.R;
 import come.geekbrains.myfragmentslessons.domain.ToolbarHolder;
-import come.geekbrains.myfragmentslessons.newer.uicard.NotesListCardFragment;
+import come.geekbrains.myfragmentslessons.newer.uicard.MyDialogFragment;
 
 public class MainActivity extends AppCompatActivity implements ToolbarHolder {
