+++ app/src/main/res/drawable/ic_divider.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+
+    android:shape="rectangle">
+    <size android:height="2dp" />
+    <solid android:color="@color/purple_500" />
+
+
+</shape>
