+++ app/src/main/res/layout/fragment_note_detail.xml
@@ -27,7 +27,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent">
 
-            <EditText
+            <TextView
                 android:id="@+id/date_creation"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
