+++ app/src/main/res/layout/item_card.xml
@@ -20,8 +20,8 @@
 
             <androidx.appcompat.widget.AppCompatImageView
                 android:id="@+id/imageView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
+                android:layout_width="@dimen/heightImage2"
+                android:layout_height="@dimen/heightImage2"
                 android:src="@drawable/birds"
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
