+++ app/src/main/res/values/dimens.xml
@@ -4,4 +4,5 @@
     <dimen name="textDescription">20sp</dimen>
     <dimen name="textTitle">42sp</dimen>
     <dimen name="heightImage">200dp</dimen>
+    <dimen name="heightImage2">100dp</dimen>
 </resources>
